[svn-r337] Changes since 19980403 ----------------------
./configure.in
	Moved setting of compiler warning switches earlier in the file.
	Turned on more warning switches to gcc.

./config/linux
	Prints a warning if the gcc version is less than 2.8.1 since
	that version has problems with register allocation for `long
	long'.

./html/Datatypes.html
	Documented sharing of data types between datasets.

./src/H5G.c
./src/H5Gpublic.h
	Implemented H5Gmove(), H5Glink() and H5Gunlink() for hard
	links.  Still have soft links to do.

./src/H5AC.c
./src/H5ACprivate.h
./src/H5D.c
./src/H5E.c
./src/H5Eprivate.h
./src/H5F.c
./src/H5Farray.c
./src/H5Fcore.c
./src/H5Ffamily.c
./src/H5Fistore.c
./src/H5Flow.c
./src/H5Fprivate.h
./src/H5Fpublic.h
./src/H5Fsec2.c
./src/H5Fstdio.c
./src/H5G.c
./src/H5Gent.c
./src/H5Gnode.c
./src/H5Gpkg.h
./src/H5Gprivate.h
./src/H5HG.c
./src/H5HL.c
./src/H5HLprivate.h
./src/H5I.c
./src/H5Iprivate.h
./src/H5MM.c
./src/H5MMprivate.h
./src/H5O.c
./src/H5Oefl.c
./src/H5Oprivate.h
./src/H5Osdspace.c
./src/H5Oshared.c
./src/H5Ostab.c
./src/H5P.c
./src/H5S.c
./src/H5Ssimp.c
./src/H5T.c
./src/H5Tconv.c
./src/H5Tprivate.h
./src/H5Tpublic.h
./src/H5V.c
./src/H5Vprivate.h
./src/H5detect.c
./src/h5ls.c
./test/cmpd_dset.c
./test/dsets.c
./test/external.c
./test/hyperslab.c
./test/iopipe.c
./test/istore.c
./test/shtype.c
./test/tstab.c
	Fixed comparisons between signed and unsigned values. Fixed
	warnings about unused function arguments.

+++ src/H5F.c
@@ -213,10 +213,10 @@ H5F_encode_length_unusual(const H5F_t *f, uint8 **p, uint8 *l)
     /*
      * For non-little-endian platforms, encode each byte in memory backwards.
      */
-    for (; i >= 0; i--, (*p)++) *(*p) = *(l + i);
+    for (/*void*/; i>=0; i--, (*p)++)*(*p) = *(l+i);
 #else
     /* platform has little-endian integers */
-    HDmemcpy(*p,l,i+1);
+    HDmemcpy(*p,l,(size_t)(i+1));
     *p+=(i+1);
 #endif
 
@@ -914,7 +914,7 @@ H5F_open(const char *name, uintn flags,
 	
 	/* nothing to check for consistency flags */
 
-	assert(p - buf == fixed_size);
+	assert((size_t)(p-buf) == fixed_size);
 
 	/* Read the variable length part of the boot block... */
 	variable_size = H5F_SIZEOF_ADDR(f) +	/*base address */
@@ -1275,13 +1275,13 @@ H5F_flush(H5F_t *f, hbool_t invalidate)
     /* update file length if necessary */
     if (!H5F_addr_defined(&(f->shared->hdf5_eof))) {
 	H5F_addr_reset(&(f->shared->hdf5_eof));
-	H5F_addr_inc(&(f->shared->hdf5_eof), p - buf);
+	H5F_addr_inc(&(f->shared->hdf5_eof), (size_t)(p-buf));
 	H5F_low_seteof(f->shared->lf, &(f->shared->hdf5_eof));
     }
     
     /* write the boot block to disk */
     if (H5F_low_write(f->shared->lf, &(f->shared->access_parms),
-		      &(f->shared->boot_addr), p-buf, buf)<0) {
+		      &(f->shared->boot_addr), (size_t)(p-buf), buf)<0) {
 	HRETURN_ERROR(H5E_IO, H5E_WRITEERROR, FAIL, "can't write header");
     }
