+++ src/H5HLprivate.h
@@ -34,7 +34,7 @@
 #define H5HL_MAGIC	"HEAP"		/*heap magic number		     */
 #define H5HL_SIZEOF_MAGIC 4
 
-#define H5HL_ALIGN(X)	(((X)+7)&~0x07) /*align on 8-byte boundary	*/
+#define H5HL_ALIGN(X)	(((X)+7)&(unsigned)(~0x07)) /*align on 8-byte boundary	*/
 
 #define H5HL_SIZEOF_HDR(F)						      \
     H5HL_ALIGN(H5HL_SIZEOF_MAGIC +	/*heap signature		*/    \
